<refactor> step3 로 패키지 변경 및 코드리뷰 수정 사항 변경, depth 2 이상 메소드 변경, 띄워쓰기, 필요없는 메소드 삭제
