+++ src/main/java/step3/Main.java
@@ -1,4 +1,4 @@
-package race;
+package step3;
 
 import java.util.List;
