+++ src/main/java/step3/MovingStrategy.java
@@ -1,4 +1,4 @@
-package race;
+package step3;
 
 public interface MovingStrategy {
     public boolean isMovable();
