+++ src/main/java/step3/OutputView.java
@@ -1,4 +1,4 @@
-package race;
+package step3;
 
 import java.util.List;
 import java.util.stream.Collectors;
