+++ src/main/java/step3/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package race;
+package step3;
 
 import java.util.Random;
