+++ src/test/java/step3/RacingCarTest.java
@@ -1,4 +1,4 @@
-package race;
+package step3;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
